third_party/WebKit/LayoutTests/http/tests/inspector/debugger/fetch-breakpoints.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../debugger-test.js"></script>
<script>

function sendRequest(url)
{
    fetch(url);
}

function test()
{
    InspectorTest.runDebuggerTestSuite([
        function testFetchBreakpoint(next)
        {
            SDK.domDebuggerManager.addXHRBreakpoint("foo", true);
            InspectorTest.waitUntilPaused(step1);
            InspectorTest.evaluateInPageWithTimeout("sendRequest('/foo?a=b')");

            function step1(callFrames)
            {
                InspectorTest.captureStackTrace(callFrames);
                InspectorTest.resumeExecution(step2);
            }

            function step2()
            {
                InspectorTest.evaluateInPage("sendRequest('/bar?a=b')", step3);
            }

            function step3()
            {
                SDK.domDebuggerManager.removeXHRBreakpoint("foo");
                InspectorTest.evaluateInPage("sendRequest('/foo?a=b')", next);
            }
        },

        function testPauseOnAnyFetch(next)
        {
            SDK.domDebuggerManager.addXHRBreakpoint("", true);
            InspectorTest.waitUntilPaused(pausedFoo);
            InspectorTest.evaluateInPageWithTimeout("sendRequest('/foo?a=b')");

            function pausedFoo(callFrames)
            {
                function resumed()
                {
                    InspectorTest.waitUntilPaused(pausedBar);
                    InspectorTest.evaluateInPage("sendRequest('/bar?a=b')");
                }
                InspectorTest.resumeExecution(resumed);
            }

            function pausedBar(callFrames)
            {
                function resumed()
                {
                    SDK.domDebuggerManager.removeXHRBreakpoint("");
                    InspectorTest.evaluateInPage("sendRequest('/baz?a=b')", next);
                }
                InspectorTest.resumeExecution(resumed);
            }
        },

        function testDisableBreakpoint(next)
        {
            SDK.domDebuggerManager.addXHRBreakpoint("", true);
            InspectorTest.waitUntilPaused(paused);
            InspectorTest.evaluateInPage("sendRequest('/foo')");

            function paused(callFrames)
            {
                function resumed()
                {
                    SDK.domDebuggerManager.toggleXHRBreakpoint("", false);
                    InspectorTest.waitUntilPaused(pausedAgain);
                    InspectorTest.evaluateInPage("sendRequest('/foo')", next);
                }
                InspectorTest.resumeExecution(resumed);
            }

            function pausedAgain(callFrames)
            {
                InspectorTest.addResult("Fail, paused again after breakpoint was removed.");
                next();
            }
        }
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests fetch() breakpoints.
</p>

</body>
</html>
